Refactor out CommitContext
